fix all models
