+++ src/transformers/models/ernie4_5/modular_ernie4_5.py
@@ -18,7 +18,7 @@ from torch import nn
 
 from ...modeling_rope_utils import dynamic_rope_update
 from ...utils import auto_docstring, can_return_tuple
-from ...utils.generic import maybe_autocast
+from ...utils.generic import maybe_autocast, no_inherit_decorator
 from ..glm.modeling_glm import rotate_half
 from ..llama.modeling_llama import (
     LlamaAttention,
@@ -90,6 +90,7 @@ class Ernie4_5MLP(LlamaMLP):
         self.down_proj = nn.Linear(self.intermediate_size, self.hidden_size, bias=config.use_bias)
 
 
+@no_inherit_decorator
 class Ernie4_5Attention(LlamaAttention):
     def __init__(self, config: Ernie4_5Config, layer_idx: int):
         super().__init__(config, layer_idx)
