+++ src/transformers/models/glm4_moe/modular_glm4_moe.py
@@ -20,7 +20,6 @@ from torch import nn
 from ...configuration_utils import PreTrainedConfig
 from ...modeling_rope_utils import RopeParameters
 from ...utils import auto_docstring, logging
-from ...utils.generic import no_inherit_decorator
 from ..cohere.modeling_cohere import CohereAttention
 from ..deepseek_v3.modeling_deepseek_v3 import (
     DeepseekV3DecoderLayer,
@@ -129,7 +128,6 @@ class Glm4MoeRotaryEmbedding(GlmRotaryEmbedding):
     pass
 
 
-@no_inherit_decorator
 class Glm4MoeAttention(CohereAttention):
     def __init__(self, config: Glm4MoeConfig, layer_idx: int | None = None):
         nn.Module.__init__(self)
