+++ src/transformers/models/glm4v_moe/modular_glm4v_moe.py
@@ -25,7 +25,7 @@ from ...modeling_outputs import MoeModelOutputWithPast
 from ...modeling_utils import ALL_ATTENTION_FUNCTIONS
 from ...processing_utils import Unpack
 from ...utils import TransformersKwargs, auto_docstring, logging
-from ...utils.generic import can_return_tuple, no_inherit_decorator
+from ...utils.generic import can_return_tuple
 from ..deepseek_v3.modeling_deepseek_v3 import DeepseekV3NaiveMoe
 from ..glm4.modeling_glm4 import Glm4Attention
 from ..glm4_moe.configuration_glm4_moe import Glm4MoeConfig
@@ -140,7 +140,6 @@ class Glm4vMoeConfig(Glm4vConfig):
     video_token_id: int = 151364
 
 
-@no_inherit_decorator
 class Glm4vMoeTextAttention(Glm4Attention):
     def __init__(self, config: Glm4vMoeTextConfig, layer_idx: int | None = None):
         super().__init__(config, layer_idx)
