+++ tests/test_modeling_common.py
@@ -5710,6 +5710,17 @@ class ModelTesterMixin:
                         with torch.no_grad():
                             _ = model(**all_inputs)
 
+    @require_kernels
+    def test_kernels_can_load_without_crashing(self):
+        """Check whether activating kernels leads to an (value) error"""
+        config, _ = self.model_tester.prepare_config_and_inputs_for_common()
+
+        for model_class in self.all_model_classes:
+            model = model_class(config)
+
+            # Using kernels should not raise a `ValueError`
+            model.use_kernels = True
+
 
 global_rng = random.Random()
